+++ app/src/pages/ImprintPage/index.tsx
@@ -1,0 +1,6 @@
+// SPDX-FileCopyrightText: OpenTalk GmbH <mail@opentalk.eu>
+//
+// SPDX-License-Identifier: EUPL-1.2
+import { ImprintPage } from './ImprintPage';
+
+export default ImprintPage;